pyproject.toml
[project]
name = "ocr-cli"
version = "1.0.0"
description = "OCR CLI tool for extracting text from images and PDFs"
requires-python = ">=3.9"
dependencies = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pypdfium2>=4.0.0",
    "pypdf>=4.0.0",
]

[project.scripts]
ocr = "ocr:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["ocr.py"]
